+++ R/RcppExports.R
@@ -13,6 +13,10 @@ sampleBinomialMaxLrr <- function(groupSizes, p, minimumEvents, sampleSize) {
     .Call('_EmpiricalCalibration_sampleBinomialMaxLrr', PACKAGE = 'EmpiricalCalibration', groupSizes, p, minimumEvents, sampleSize)
 }
 
+samplePoissonRegressionMaxLrr <- function(groupSizes, z, minimumEvents, sampleSize) {
+    .Call('_EmpiricalCalibration_samplePoissonRegressionMaxLrr', PACKAGE = 'EmpiricalCalibration', groupSizes, z, minimumEvents, sampleSize)
+}
+
 logLikelihoodNull <- function(theta, logRr, seLogRr) {
     .Call('_EmpiricalCalibration_logLikelihoodNull', PACKAGE = 'EmpiricalCalibration', theta, logRr, seLogRr)
 }