+++ NEWS.md
@@ -3,11 +3,14 @@ EmpiricalCalibration 3.0.1
 
 Changes
 
-1. Adding the [compareEase()] function to compare EASE of correlated sets of estimates.
+1. Adding the `compareEase()` function to compare EASE of correlated sets of estimates.
+
+2. Adding the `computeCvPoissonRegression()` function to compute critical values for Poisson regressions. 
 
 Bugfixes
 
 1. Fixing `computeTraditionalP()` when using vectors as input.
+
 
 EmpiricalCalibration 3.0.0
 ==========================